requirements.txt
streamlit
langchain
langchain-huggingface
huggingface-hub
python-dotenv
